requirements.txt
opencv-python==4.8.1.78
flask==3.0.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
werkzeug==3.0.1
gunicorn==21.2.0